added in scrolling to the game chat box. Now only auto scrolls when the player has sent a message or is scrolled at the bottom already. If they scroll up, it doesn't scroll them down on every message received.
