+++ application/__init__.py
@@ -20,7 +20,7 @@ from . import models
 # Game sockets import game mechanics and socketio listeners.
 from .controllers.games import chase_the_ace_gameplay
 from .controllers.games import chase_the_ace_redirects
-from .controllers import game_chat
+from .controllers import chat
 
 
 def create_app():
